pyproject.toml
[build-system]
# build the package with [flit](https://flit.readthedocs.io)
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

[project]
# See https://www.python.org/dev/peps/pep-0621/
name = "aiida-skeaf"
dynamic = ["version"]  # read from aiida_skeaf/__init__.py
description = "AiiDA plugin for the Supercell K-space Extremal Area Finder (SKEAF) code"
authors = [{name = "Junfeng Qiao", email = "qiaojunfeng@outlook.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Development Status :: 3 - Alpha",
    "Framework :: AiiDA"
]
keywords = ["aiida", "plugin"]
requires-python = ">=3.7"
dependencies = [
    "aiida-core>=1.6.4,<3",
    "voluptuous"
]

[project.urls]
Source = "https://github.com/qiaojunfeng/aiida-skeaf"

[project.optional-dependencies]
testing = [
    "pgtest~=1.3.1",
    "wheel~=0.31",
    "coverage[toml]",
    "pytest~=6.0",
    "pytest-cov",
    "pytest-regressions"
]
pre-commit = [
    "pre-commit~=2.2",
    "pylint>=2.5.0,<2.9"
]
docs = [
    "sphinx",
    "sphinxcontrib-contentui",
    "sphinxcontrib-details-directive",
    "furo",
    "markupsafe<2.1"
]
analysis = [
    "matplotlib"
]

# [project.entry-points."aiida.data"]
# "skeaf" = "aiida_skeaf.data:SkeafParameters"

[project.entry-points."aiida.calculations"]
"skeaf.skeaf" = "aiida_skeaf.calculations:SkeafCalculation"
"skeaf.create_bxsf_from_wannier90" = "aiida_skeaf.calculations.functions:create_bxsf_from_wannier90"
"skeaf.create_bxsf_from_file" = "aiida_skeaf.calculations.functions:create_bxsf_from_file"

[project.entry-points."aiida.parsers"]
"skeaf" = "aiida_skeaf.parsers:SkeafParser"

# [project.entry-points."aiida.cmdline.data"]
# "skeaf" = "aiida_skeaf.cli:data_cli"

[tool.flit.module]
name = "aiida_skeaf"

[tool.pylint.format]
max-line-length = 125

[tool.pylint.messages_control]
disable = [
    "too-many-ancestors",
    "invalid-name",
    "duplicate-code",
    "import-outside-toplevel",
    # black compatibility
    "C0330",
    "C0326",
]

[tool.pytest.ini_options]
# Configuration for [pytest](https://docs.pytest.org)
python_files = "test_*.py example_*.py"
filterwarnings = [
    "ignore::DeprecationWarning:aiida:",
    "ignore::DeprecationWarning:plumpy:",
    "ignore::DeprecationWarning:django:",
    "ignore::DeprecationWarning:yaml:",
]

[tool.coverage.run]
# Configuration of [coverage.py](https://coverage.readthedocs.io)
# reporting which lines of your plugin are covered by tests
source=["aiida_skeaf"]

[tool.isort]
# Configuration of [isort](https://isort.readthedocs.io)
line_length = 120
force_sort_within_sections = true
sections = ['FUTURE', 'STDLIB', 'THIRDPARTY', 'AIIDA', 'FIRSTPARTY', 'LOCALFOLDER']
known_aiida = ['aiida']

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py38

[testenv]
usedevelop=True

[testenv:py{37,38,39,310}]
description = Run the test suite against a python version
extras = testing
commands = pytest {posargs}

[testenv:pre-commit]
description = Run the pre-commit checks
extras = pre-commit
commands = pre-commit run {posargs}

[testenv:docs]
description = Build the documentation
extras = docs
commands = sphinx-build -nW --keep-going -b html {posargs} docs/source docs/build/html
commands_post = echo "open file://{toxinidir}/docs/build/html/index.html"
"""
